- Refactored common things into AbstractSchemaParserTest - Implemented VastXsdSchemaParserTest
